Qt: Check screen bounds when restoring window positions
Ensure that the geometry is not off-screen.

It won't work on Wankland, and apparently doesn't support multiple monitors on X11.

So I'm just going to disable the whole thing on Linux, not just because my care
levels are practically non-existant, but also because I don't want to deal with
people moaning that their window manager's behavior is causing positions to
revert to the primary monitor. Just yolo it and hope for the best....